lib/router - refactor, impl route module prefetching
