+++ packages/lib/src/router/context.ts
@@ -8,6 +8,7 @@ export interface FileRouterContextType {
    * The current router state
    */
   state: RouterState
+
   /**
    * Navigate to a new route, optionally replacing the current route
    * in the history stack or triggering a view transition
@@ -16,10 +17,17 @@ export interface FileRouterContextType {
     path: string,
     options?: { replace?: boolean; transition?: boolean }
   ) => Promise<void>
+
+  /**
+   * Prefetch a route module and its dependencies to be loaded in the background
+   */
+  prefetchRouteModules: (path: string) => void
+
   /**
    * Reload the current route, optionally triggering a view transition
    */
   reload: (options?: { transition?: boolean }) => Promise<void>
+
   /**
    * Set the current query parameters
    */
